@@ -1,3 +1,3 @@
 ﻿namespace quiz_management_system.Application.Dtos;
 
-public record AuthenticatedUser(string Id, string Email, string FullName, string Role);
+public record AuthenticatedUser(Guid Id, string Email, string FullName, string Role);
