@@ -4,7 +4,7 @@ using quiz_management_system.Domain.Common.ResultPattern.Result;
 namespace quiz_management_system.Application.Features.ResetPassword
 {
     public sealed record ResetPasswordWithCodeCommand(
-        string UserId,
+        Guid UserId,
         string Code,
         string NewPassword,
         string UserIpAddress
