@@ -2,7 +2,7 @@
 
 public sealed record ResetPasswordRequest
 {
-    public required string UserId { get; init; }
+    public required Guid UserId { get; init; }
     public required string Code { get; init; }
     public required string NewPassword { get; init; }
 }
