@@ -2,4 +2,4 @@
 
 namespace quiz_management_system.Domain.Events;
 
-public record ResetPasswordEvent(string userId, string Email, string FullName, string Role) : DomainEvent;
+public record ResetPasswordEvent(Guid userId, string Email, string FullName, string Role) : DomainEvent;
