@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using quiz_management_system.Domain.Users.Abstraction;
 using quiz_management_system.Domain.Users.StudentsFolder;
 
 namespace quiz_management_system.Infrastructure.Data.Configurations;
@@ -10,22 +11,19 @@ public sealed class StudentConfiguration : IEntityTypeConfiguration<Student>
     {
         builder.ToTable("Students");
 
-        builder.HasKey(x => x.Id);
 
-        builder.Property(x => x.Id)
-               .HasColumnType("uuid")
-               .IsRequired();
-
-        builder.Property(x => x.FullName)
-               .IsRequired()
-               .HasMaxLength(200);
-
-        builder.Property(x => x.Email)
-               .IsRequired()
-               .HasMaxLength(256);
+        builder.HasBaseType<DomainUser>();
 
         builder.Property(x => x.AcademicYearId)
                .HasColumnType("uuid")
                .IsRequired();
+
+        builder.HasOne(x => x.AcademicYear)
+               .WithMany()
+               .HasForeignKey(x => x.AcademicYearId)
+               .OnDelete(DeleteBehavior.Restrict);
+
+        builder.Navigation(x => x.AcademicYear)
+               .UsePropertyAccessMode(PropertyAccessMode.Property);
     }
 }
