@@ -1,45 +1,60 @@
-﻿using Microsoft.AspNetCore.Identity.UI.Services;
+﻿using MailKit.Net.Smtp;
+using MailKit.Security;
+using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using MimeKit;
 using quiz_management_system.Infrastructure.Email;
-using Resend;
 
-public sealed class ResendEmailSender : IEmailSender
+public sealed class EmailSender : IEmailSender
 {
-    private readonly ResendSettings _settings;
-    private readonly ILogger<ResendEmailSender> _logger;
-    private readonly IResend _resend;
+    private readonly EmailSettings _settings;
+    private readonly ILogger<EmailSender> _logger;
 
-    public ResendEmailSender(
-        IOptions<ResendSettings> settings,
-        ILogger<ResendEmailSender> logger,
-        IResend resend)
+    public EmailSender(
+        IOptions<EmailSettings> settings,
+        ILogger<EmailSender> logger)
     {
         _settings = settings.Value;
         _logger = logger;
-        _resend = resend;
     }
 
     public async Task SendEmailAsync(string email, string subject, string htmlMessage)
     {
-        var message = new EmailMessage
+        var message = new MimeMessage();
+
+        message.From.Add(new MailboxAddress(_settings.FromName, _settings.FromAddress));
+        message.To.Add(MailboxAddress.Parse(email));
+        message.Subject = subject;
+
+        message.Body = new TextPart("html")
         {
-            From = _settings.From,
-            Subject = subject,
-            HtmlBody = htmlMessage
+            Text = htmlMessage
         };
 
-        message.To.Add(email);
+        using var client = new SmtpClient();
 
         try
         {
-            var result = await _resend.EmailSendAsync(message);
-            _logger.LogInformation("Email sent via Resend to {Recipient}. Id={Id}", email, result.Content);
+            await client.ConnectAsync(
+                _settings.Host,
+                _settings.Port,
+                _settings.UseSsl ? SecureSocketOptions.StartTls : SecureSocketOptions.Auto);
+
+            if (!string.IsNullOrWhiteSpace(_settings.Username))
+            {
+                await client.AuthenticateAsync(
+                    _settings.Username,
+                    _settings.Password);
+            }
+
+            await client.SendAsync(message);
         }
-        catch (Exception ex)
+        finally
         {
-            _logger.LogError(ex, "Error sending email via Resend");
-            throw;
+            await client.DisconnectAsync(true);
         }
+
+        _logger.LogInformation("Email sent to {Recipient}", email);
     }
 }
